Add instructions for using v7
We've just added the v7 CLI to our release.  Added instructions for selecting the CLI required by consumers.

While it's possible to use the `cf-cli-release` to lay down both a v6 and v7 version of the CLI on the same VM, we don't believe this is a common use case.  Please let us know if you need this supported and documented.